test(native): co-locate onboarding and earn tests
