+++ suite-native/module-earn/src/components/ClaimTransactionDataReviewStepList.test.tsx
@@ -3,21 +3,20 @@ import { getTranslation } from '@suite-native/intl';
 import { renderWithStoreProvider } from '@suite-native/test-utils-store';
 import { type ReviewSummaryOutput } from '@suite-native/transaction-management';
 
-import { UnstakeTransactionDataReviewStepList } from '../UnstakeTransactionDataReviewStepList';
+import { ClaimTransactionDataReviewStepList } from './ClaimTransactionDataReviewStepList';
 
-const mockUnstakeOutputItem = jest.fn();
+const mockClaimOutputItem = jest.fn();
 const mockEarnSummaryOutputItem = jest.fn();
 
 let mockIsTransactionAlreadySigned: boolean;
 let mockSummaryOutput: ReviewSummaryOutput | null;
 let mockAccountNetworkSymbol: string | null;
-let mockPrecomposed: { fee: string; solanaTxMeta?: { deviceAmountLamports: string } };
 
 const accountKey = mockAccountKey({ symbol: 'eth', descriptor: 'ethAccount' });
 
 jest.mock('@react-navigation/native', () => ({
     ...jest.requireActual('@react-navigation/native'),
-    useRoute: () => ({ params: { accountKey, amount: '1' } }),
+    useRoute: () => ({ params: { accountKey } }),
 }));
 
 jest.mock('@suite-native/transaction-management', () => ({
@@ -31,15 +30,20 @@ jest.mock('@suite-common/wallet-core', () => ({
     selectAccountNetworkSymbol: () => mockAccountNetworkSymbol,
 }));
 
-jest.mock('../UnstakeOutputItem', () => ({
-    UnstakeOutputItem: (props: { outputState?: string }) => {
-        mockUnstakeOutputItem(props);
+jest.mock('@suite-native/staking', () => ({
+    ...jest.requireActual('@suite-native/staking'),
+    selectClaimableAmountByAccountKey: () => '5',
+}));
+
+jest.mock('./ClaimOutputItem', () => ({
+    ClaimOutputItem: (props: { outputState?: string }) => {
+        mockClaimOutputItem(props);
 
         return null;
     },
 }));
 
-jest.mock('../EarnSummaryOutputItem', () => ({
+jest.mock('./EarnSummaryOutputItem', () => ({
     EarnSummaryOutputItem: (props: { outputState?: string; fee?: string }) => {
         mockEarnSummaryOutputItem(props);
 
@@ -47,45 +51,35 @@ jest.mock('../EarnSummaryOutputItem', () => ({
     },
 }));
 
-jest.mock('../../hooks/useEarnSelectedPrecomposedTransaction', () => ({
-    useEarnSelectedPrecomposedTransaction: () => mockPrecomposed,
+jest.mock('../hooks/useEarnSelectedPrecomposedTransaction', () => ({
+    useEarnSelectedPrecomposedTransaction: () => ({ fee: '21000', totalSpent: '5' }),
 }));
 
-const renderStepList = () => renderWithStoreProvider(<UnstakeTransactionDataReviewStepList />);
+const renderStepList = () => renderWithStoreProvider(<ClaimTransactionDataReviewStepList />);
 
-describe('UnstakeTransactionDataReviewStepList', () => {
+describe('ClaimTransactionDataReviewStepList', () => {
     beforeEach(() => {
         jest.clearAllMocks();
         mockIsTransactionAlreadySigned = false;
         mockSummaryOutput = null;
         mockAccountNetworkSymbol = 'eth';
-        mockPrecomposed = { fee: '21000' };
     });
 
-    it('renders nothing until the account network symbol is known', () => {
-        mockAccountNetworkSymbol = null;
-
-        renderStepList();
-
-        expect(mockUnstakeOutputItem).not.toHaveBeenCalled();
-        expect(mockEarnSummaryOutputItem).not.toHaveBeenCalled();
-    });
-
-    it('renders the unstake and summary cards without a Next button', () => {
+    it('renders the claim and summary cards without a Next button', () => {
         const { queryByText, queryByTestId } = renderStepList();
 
-        expect(mockUnstakeOutputItem).toHaveBeenCalledTimes(1);
+        expect(mockClaimOutputItem).toHaveBeenCalledTimes(1);
         expect(mockEarnSummaryOutputItem).toHaveBeenCalledTimes(1);
         expect(queryByText(getTranslation('generic.buttons.next'))).toBeNull();
-        expect(queryByTestId('@earn/unstake-review-continue')).toBeNull();
+        expect(queryByTestId('@earn/claim-review-continue')).toBeNull();
     });
 
-    it('keeps the unstake step active and the summary hidden until the device confirms the outputs', () => {
+    it('keeps the claim step active and the summary hidden until the device confirms the outputs', () => {
         mockSummaryOutput = null;
 
         renderStepList();
 
-        expect(mockUnstakeOutputItem).toHaveBeenCalledWith(
+        expect(mockClaimOutputItem).toHaveBeenCalledWith(
             expect.objectContaining({ outputState: 'active' }),
         );
         expect(mockEarnSummaryOutputItem).toHaveBeenCalledWith(
@@ -93,12 +87,12 @@ describe('UnstakeTransactionDataReviewStepList', () => {
         );
     });
 
-    it('marks the unstake step done and activates the summary once all device outputs are confirmed', () => {
-        mockSummaryOutput = { state: 'active', totalSpent: '1', fee: '21000' };
+    it('marks the claim step done and activates the summary once all device outputs are confirmed', () => {
+        mockSummaryOutput = { state: 'active', totalSpent: '5', fee: '21000' };
 
         renderStepList();
 
-        expect(mockUnstakeOutputItem).toHaveBeenCalledWith(
+        expect(mockClaimOutputItem).toHaveBeenCalledWith(
             expect.objectContaining({ outputState: 'success' }),
         );
         expect(mockEarnSummaryOutputItem).toHaveBeenCalledWith(
@@ -108,11 +102,11 @@ describe('UnstakeTransactionDataReviewStepList', () => {
 
     it('marks both steps done and hides the sliding footer once the transaction is signed', () => {
         mockIsTransactionAlreadySigned = true;
-        mockSummaryOutput = { state: 'success', totalSpent: '1', fee: '21000' };
+        mockSummaryOutput = { state: 'success', totalSpent: '5', fee: '21000' };
 
         const { queryByTestId } = renderStepList();
 
-        expect(mockUnstakeOutputItem).toHaveBeenCalledWith(
+        expect(mockClaimOutputItem).toHaveBeenCalledWith(
             expect.objectContaining({ outputState: 'success' }),
         );
         expect(mockEarnSummaryOutputItem).toHaveBeenCalledWith(
@@ -125,14 +119,5 @@ describe('UnstakeTransactionDataReviewStepList', () => {
         const { getByTestId } = renderStepList();
 
         expect(getByTestId('sliding-footer-overlay')).toBeOnTheScreen();
-    });
-
-    it('passes the amount in base units and the unstake stake type to the summary', () => {
-        renderStepList();
-
-        // Route amount '1' ETH -> 1e18 wei. The summary decides itself whether to show it.
-        expect(mockEarnSummaryOutputItem).toHaveBeenCalledWith(
-            expect.objectContaining({ amount: '1000000000000000000', stakeType: 'unstake' }),
-        );
     });
 });
