+++ suite-native/module-earn/src/unstakeFormSchema.test.ts
@@ -2,7 +2,7 @@ import { type Account } from '@suite-common/wallet-types';
 import { type SolanaStakingAccount } from '@trezor/blockchain-link-types';
 import { StakeState } from '@trezor/network-solana/constants';
 
-import { type UnstakeFormContext, unstakeFormValidationSchema } from '../unstakeFormSchema';
+import { type UnstakeFormContext, unstakeFormValidationSchema } from './unstakeFormSchema';
 
 const SOL = 1_000_000_000;
